修复溢出BUG

+++ src/main/java/com/piggest/minecraft/bukkit/utils/Radio.java
@@ -83,9 +83,9 @@ public class Radio {
 	public static double get_power_at(Location source_location, double radiant_power, int central_freq, Location loc) {
 		double distance = 1;
 		if (source_location.getWorld() != loc.getWorld()) {
-			int x_distance = source_location.getBlockX() - loc.getBlockX();
-			int y_distance = source_location.getBlockY() - loc.getBlockY();
-			int z_distance = source_location.getBlockZ() - loc.getBlockZ();
+			double x_distance = source_location.getBlockX() - loc.getBlockX();
+			double y_distance = source_location.getBlockY() - loc.getBlockY();
+			double z_distance = source_location.getBlockZ() - loc.getBlockZ();
 			distance = 256 + Math.sqrt(x_distance * x_distance + y_distance * y_distance + z_distance * z_distance);
 		} else {
 			distance = source_location.distance(loc);
